+++ pyaud/_wraps.py
@@ -14,8 +14,8 @@ import spall.exceptions as sp_exceptions
 
 from . import _data
 from ._cache import FileCacher as _FileCacher
-from ._environ import environ as _environ
 from ._indexing import files as _files
+from ._locations import AppFiles as _AppFiles
 from ._objects import BasePlugin as _BasePlugin
 from ._utils import colors as _colors
 
@@ -83,12 +83,14 @@ class ClassDecorator:
     ``__call__`` method.
 
     :param cls: The class whose ``__call__`` method will be wrapped.
+    :param app_files: App file locations object.
     """
 
-    DURATIONS = "durations.json"
-
-    def __init__(self, cls: _t.Type[_BasePlugin]) -> None:
+    def __init__(
+        self, cls: _t.Type[_BasePlugin], app_files: _AppFiles
+    ) -> None:
         self._cls = cls
+        self._app_files = app_files
 
     def time(self, func: _t.Callable[..., int]) -> _t.Callable[..., int]:
         """Wrap ``__call__`` with a timer.
@@ -101,11 +103,11 @@ class ClassDecorator:
         def _wrapper(*args: str, **kwargs: bool) -> int:
             repo = _Path.cwd().name
             with _data.record.track(
-                repo, self._cls, _environ.DATADIR / "durations.json"
+                repo, self._cls, self._app_files.durations_file
             ) as time_keeper:
                 returncode = func(*args, **kwargs)
 
-            _data.write(_data.record, _environ.DATADIR / _data.DURATIONS)
+            _data.write(_data.record, self._app_files.durations_file)
             logged_time = "{}: Execution time: {}s; Average time: {}s".format(
                 self._cls.__name__,
                 time_keeper.elapsed(),
@@ -128,7 +130,9 @@ class ClassDecorator:
 
         @_functools.wraps(func)
         def _wrapper(*args: str, **kwargs: bool) -> int:
-            _file_cacher = _FileCacher(self._cls, func, *args, **kwargs)
+            _file_cacher = _FileCacher(
+                self._cls, func, self._app_files, *args, **kwargs
+            )
             return _file_cacher.files(func, *args, **kwargs)
 
         return _wrapper