+++ pyaud/config.py
@@ -61,10 +61,10 @@ import toml.decoder as _toml_decoder
 import toml.encoder as _toml_encoder
 
 from . import _typing as _pt
-from ._environ import environ as _environ
+from ._locations import NAME as _NAME
+from ._locations import AppFiles as _AppFiles
 from ._objects import MutableMapping as _MutableMapping
 
-CONFIGDIR = _Path(_appdirs.user_config_dir(_environ.NAME))
 DEFAULT_CONFIG: _t.Dict[str, _t.Any] = dict(
     clean={"exclude": ["*.egg*", ".mypy_cache", ".env", "instance"]},
     logging={
@@ -82,8 +82,7 @@ DEFAULT_CONFIG: _t.Dict[str, _t.Any] = dict(
                 "when": "d",
                 "backupCount": 60,
                 "filename": str(
-                    _Path(_appdirs.user_log_dir(_environ.NAME))
-                    / f"{_environ.NAME}.log"
+                    _Path(_appdirs.user_log_dir(_NAME)) / f"{_NAME}.log"
                 ),
             }
         },
@@ -107,8 +106,6 @@ DEFAULT_CONFIG: _t.Dict[str, _t.Any] = dict(
     },
     addopts=["timed"],
 )
-
-_TOMLFILE = f"{_environ.NAME}.toml"
 
 
 class _TomlArrayEncoder(_toml_encoder.TomlEncoder):
@@ -222,7 +219,7 @@ class TempEnvVar:
                 self._obj[key] = self._default[key]
 
 
-def configure_global() -> None:
+def configure_global(app_files: _AppFiles) -> None:
     """Setup object with default config settings.
 
     Create config file with default config settings if one does not
@@ -230,41 +227,48 @@ def configure_global() -> None:
 
     Load base config file which may, or may not, still have the default
     settings configured.
+
+    :param app_files: App file locations object.
     """
-    configfile = CONFIGDIR / _TOMLFILE
-    backupfile = CONFIGDIR / f".{_TOMLFILE}.bak"
     default_config = _copy.deepcopy(DEFAULT_CONFIG)
     toml.update(default_config)
-    if configfile.is_file():
+    if app_files.global_config_file.is_file():
         while True:
-            with open(configfile, encoding="utf-8") as fin:
+            with open(app_files.global_config_file, encoding="utf-8") as fin:
                 try:
                     toml.load(fin)
-                    _shutil.copyfile(configfile, backupfile)
+                    _shutil.copyfile(
+                        app_files.global_config_file,
+                        app_files.global_config_file_backup,
+                    )
                     break
 
                 except _toml_decoder.TomlDecodeError:
-                    if backupfile.is_file():
-                        _os.rename(backupfile, configfile)
+                    if app_files.global_config_file_backup.is_file():
+                        _os.rename(
+                            app_files.global_config_file_backup,
+                            app_files.global_config_file,
+                        )
                     else:
                         break
 
-    CONFIGDIR.mkdir(exist_ok=True, parents=True)
-    with open(configfile, "w", encoding="utf-8") as fout:
+    with open(app_files.global_config_file, "w", encoding="utf-8") as fout:
         toml.dump(fout)
 
 
-def load_config(opt: _t.Optional[_t.Union[str, _os.PathLike]] = None) -> None:
+def load_config(
+    app_files: _AppFiles, opt: _t.Optional[_t.Union[str, _os.PathLike]] = None
+) -> None:
     """Load configs in order, each one overriding the previous.
 
+    :param app_files: App file locations object.
     :param opt: Optional extra path which will override all others.
     """
-    rcfile = f".{_environ.NAME}rc"
     files = [
-        CONFIGDIR / _TOMLFILE,
-        _Path.home() / rcfile,
-        _Path.cwd() / rcfile,
-        _Path.cwd() / "pyproject.toml",
+        app_files.global_config_file,
+        app_files.home_config_file,
+        app_files.project_config_file,
+        app_files.pyproject_toml,
     ]
     if opt is not None:
         files.append(_Path(opt))
@@ -272,7 +276,7 @@ def load_config(opt: _t.Optional[_t.Union[str, _os.PathLike]] = None) -> None:
     for file in files:
         if file.is_file():
             with open(file, encoding="utf-8") as fin:
-                toml.load(fin, "tool", _environ.NAME)
+                toml.load(fin, "tool", _NAME)
 
 
 def _extract_logger(default: _t.Dict[str, _t.Any]) -> _logging.Logger: