+++ tests/cache_test.py
@@ -18,6 +18,7 @@ from . import (
     FILES,
     INIT,
     REPO,
+    AppFiles,
     CacheDict,
     CacheUnion,
     ClsDict,
@@ -66,10 +67,13 @@ def test_no_cache(monkeypatch: pytest.MonkeyPatch, main: MockMainType) -> None:
     assert save_cache.was_called() is False
 
 
-def test_remove_matched_files(monkeypatch: pytest.MonkeyPatch) -> None:
+def test_remove_matched_files(
+    monkeypatch: pytest.MonkeyPatch, app_files: AppFiles
+) -> None:
     """Test that correct files are removed for matching md5 hashes.
 
     :param monkeypatch: Mock patch environment and attributes.
+    :param app_files: App file locations object.
     """
     remove = Tracker()
     pyaud.plugins._files.append(Path.cwd() / REPO)
@@ -77,7 +81,9 @@ def test_remove_matched_files(monkeypatch: pytest.MonkeyPatch) -> None:
     monkeypatch.setattr("pyaud._cache._Path.read_bytes", lambda *_: b"")
     monkeypatch.setattr("pyaud._cache.HashMapping.match_file", lambda *_: True)
     # noinspection PyUnresolvedReferences
-    class_decorator = pyaud._wraps.ClassDecorator(MockCachedPluginType)
+    class_decorator = pyaud._wraps.ClassDecorator(
+        MockCachedPluginType, app_files
+    )
     MockCachedPluginType.__call__ = class_decorator.files(  # type: ignore
         MockCachedPluginType.__call__
     )
@@ -180,10 +186,6 @@ class TestCacheStrategy:
     def _success_msg(i: int) -> str:
         return f"Success: no issues found in {i} source files"
 
-    @property
-    def _cache_file(self) -> Path:
-        return pyaud.environ.CACHEDIR / "files.json"
-
     @staticmethod
     def _fmt(o: t.Dict[Path, str]) -> FileHashDict:
         return {str(k.relative_to(Path.cwd())): v for k, v in o.items()}
@@ -193,13 +195,17 @@ class TestCacheStrategy:
         "unpatch_hash_mapping_hash_files", "unpatch_hash_mapping_match_file"
     )
     def test_cache(
-        self, monkeypatch: pytest.MonkeyPatch, nocolorcapsys: NoColorCapsys
+        self,
+        monkeypatch: pytest.MonkeyPatch,
+        nocolorcapsys: NoColorCapsys,
+        app_files: AppFiles,
     ) -> None:
         """Test cache strategy.
 
         :param monkeypatch: Mock patch environment and attributes.
         :param nocolorcapsys: Capture system output while stripping ANSI
             color codes.
+        :param app_files: App file locations object.
         """
         #: PATHS
         p = (
@@ -247,7 +253,7 @@ class TestCacheStrategy:
         #: last process called.
         i1 = self._get_instance(monkeypatch, 1, 0)
         i1()
-        o = json.loads(self._cache_file.read_text())
+        o = json.loads(app_files.cache_file.read_text())
         assert self._success_msg(len(f)) in nocolorcapsys.stdout()
         assert self._cls_in_commit(o, 0, 1, 0)
         assert self._d_eq(self._idx(o, 0, 1, 0), self._fmt(f))
@@ -258,7 +264,7 @@ class TestCacheStrategy:
         #: Test that a success message notifies user that no changes
         #: have been made, so no process needed to be run.
         i1()
-        o = json.loads(self._cache_file.read_text())
+        o = json.loads(app_files.cache_file.read_text())
         assert self.NO_CHANGE_MSG in nocolorcapsys.stdout()
         assert self._cls_in_commit(o, 0, 1, 0)
         assert self._d_eq(self._idx(o, 0, 1), self._idx(o, 0, 0))
@@ -269,7 +275,7 @@ class TestCacheStrategy:
         #: have been made, so only a partial process needed to be run.
         f[p[0]] = f[p[0]][::-1]
         i1()
-        o = json.loads(self._cache_file.read_text())
+        o = json.loads(app_files.cache_file.read_text())
         assert self._success_msg(1) in nocolorcapsys.stdout()
         assert self._cls_in_commit(o, 0, 1, 0)
         assert self._d_eq(self._idx(o, 0, 1), self._idx(o, 0, 0))
@@ -284,7 +290,7 @@ class TestCacheStrategy:
         #: the last process that ran.
         i2 = self._get_instance(monkeypatch, 2, 0)
         i2()
-        o = json.loads(self._cache_file.read_text())
+        o = json.loads(app_files.cache_file.read_text())
         assert self.NO_CHANGE_MSG in nocolorcapsys.stdout()
         assert self._cls_in_commit(o, 0, 1, 0)
         assert self._d_eq(self._idx(o, 0, 1), self._idx(o, 0, 0))
@@ -305,7 +311,7 @@ class TestCacheStrategy:
         #: it is still the last process that ran.
         i3 = self._get_instance(monkeypatch, 1, 1)
         i3()
-        o = json.loads(self._cache_file.read_text())
+        o = json.loads(app_files.cache_file.read_text())
         assert self._success_msg(len(f)) in nocolorcapsys.stdout()
         assert self._cls_in_commit(o, 0, 1, 0)
         assert self._cls_in_commit(o, 0, 1, 1)
@@ -330,7 +336,7 @@ class TestCacheStrategy:
         #: process that ran.
         i4 = self._get_instance(monkeypatch, 3, 1)
         i4()
-        o = json.loads(self._cache_file.read_text())
+        o = json.loads(app_files.cache_file.read_text())
         assert self.NO_CHANGE_MSG in nocolorcapsys.stdout()
         assert self._cls_in_commit(o, 0, 1, 0)
         assert self._cls_in_commit(o, 0, 1, 1)
@@ -349,7 +355,7 @@ class TestCacheStrategy:
         #: called.
         i5 = self._get_instance(monkeypatch, 3, 1, clean=False)
         i5()
-        o = json.loads(self._cache_file.read_text())
+        o = json.loads(app_files.cache_file.read_text())
         assert self.NO_CHANGE_MSG in nocolorcapsys.stdout()
         assert self._cls_in_commit(o, 0, 1, 0)
         assert self._cls_in_commit(o, 0, 1, 1)
@@ -375,7 +381,7 @@ class TestCacheStrategy:
         i6 = self._get_instance(monkeypatch, 1, 0, cache_all=True)
         f[p[1]] = f[p[1]][::-1]
         i6()
-        o = json.loads(self._cache_file.read_text())
+        o = json.loads(app_files.cache_file.read_text())
         assert self._success_msg(len(f)) in nocolorcapsys.stdout()
         assert self._cls_in_commit(o, 0, 1, 0)
         assert self._d_eq(self._idx(o, 0, 1), self._idx(o, 0, 0))