+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/pyaud/compare/v3.13.5...HEAD)
 ------------------------------------------------------------------------
 ### Removed
+- Removes env config for app dirs and files
 - Removes `Write` plugin
 
 [3.13.5](https://github.com/jshwi/pyaud/releases/tag/v3.13.5) - 2022-05-05